+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.1.0] - 2025-10-26
-- initial version
+- initial version (Time Length Compressed Circular Buffer)